29. Divide Two Integers

+++ cpp/021-030/Divide Two Integers.cpp
@@ -32,7 +32,7 @@ public:
             new_divisor /= 2;
         }
         
-        return (long)sign * result;
+        return sign * result;
     }
 };
